изменила тип данных

+++ compl.py
@@ -1,21 +1,16 @@
-
 def make_complex():
     while True:
         a = (input("Введите действительную часть числа: "))
         if not (a.isdigit()):
             print('Ошибка. Повторите ввод')
             continue
-        a = int(a)
+        a = float(a)
         break
     while True:
         b = (input("Введите мнимую часть числа: "))
         if not (b.isdigit()):
             print('Ошибка. Повторите ввод')
             continue
-        b = int(b)
+        b = float(b)
         break
     return complex(a, b)
-
-
-
-
