+++ user_interface.py
@@ -3,4 +3,4 @@ def view_data(data):
     print(data)
 
 def get_value():
-    return int(input('value = '))
+    return float(input('value = '))
